libc0: Use |unsigned int| for |size_t|

+++ src/libc0/include/sys/types.h
@@ -19,6 +19,6 @@
 
 #pragma once
 
-typedef unsigned long size_t;
+typedef unsigned int  size_t;
 typedef   signed long ssize_t;
 typedef   signed long off_t;
